# docs/openapi/dispatch-service.openapi.yaml
openapi: 3.0.3
info:
  title: Dispatch Service API
  version: 0.1.0
  description: API para terminales, rutas, tipos de bus, buses, layouts de asientos y salidas programadas.
servers:
  - url: https://dispatch-service-dev.example.com/api/v1/dispatch
    description: Dev placeholder
security:
  - bearerAuth: []
...
paths:
...
  /terminals/{terminalId}:
    get:
      tags: [terminals]
      summary: Obtener terminal.
      parameters:
        - $ref: "#/components/parameters/TerminalId"
      responses:
        "200":
          description: Terminal encontrada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Terminal"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [terminals]
      summary: Actualizar terminal.
      parameters:
        - $ref: "#/components/parameters/TerminalId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTerminalRequest"
      responses:
        "200":
          description: Terminal actualizada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Terminal"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Terminal duplicada o en uso.
    delete:
      tags: [terminals]
      summary: Desactivar terminal.
      parameters:
        - $ref: "#/components/parameters/TerminalId"
      responses:
        "204":
          description: Terminal desactivada.
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Terminal en uso por rutas o buses activos.
...
  /bus-types/{busTypeId}:
    get:
      tags: [bus-types]
      summary: Obtener tipo de bus.
      parameters:
        - $ref: "#/components/parameters/BusTypeId"
      responses:
        "200":
          description: Tipo de bus encontrado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusType"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [bus-types]
      summary: Actualizar tipo de bus.
      parameters:
        - $ref: "#/components/parameters/BusTypeId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBusTypeRequest"
      responses:
        "200":
          description: Tipo de bus actualizado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusType"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Tipo de bus duplicado.
    delete:
      tags: [bus-types]
      summary: Desactivar tipo de bus.
      parameters:
        - $ref: "#/components/parameters/BusTypeId"
      responses:
        "204":
          description: Tipo de bus desactivado.
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Tipo de bus en uso por buses activos.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    TerminalId:
      name: terminalId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    BusTypeId:
      name: busTypeId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Solicitud invalida.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Recurso no encontrado.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        correlation_id:
          type: string
          format: uuid
...
    Terminal:
      type: object
      required: [id, name, active]
      properties:
        id:
          type: string
          format: uuid
        legacy_id:
          type: integer
        local_code:
          type: string
        name:
          type: string
        manager_name:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email
        active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    UpdateTerminalRequest:
      type: object
      properties:
        legacy_id:
          type: integer
        local_code:
          type: string
        name:
          type: string
        manager_name:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
          format: email
        active:
          type: boolean
...
    BusType:
      type: object
      required: [id, name, active]
      properties:
        id:
          type: string
          format: uuid
        legacy_id:
          type: integer
        name:
          type: string
        description:
          type: string
        active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    UpdateBusTypeRequest:
      type: object
      properties:
        legacy_id:
          type: integer
        name:
          type: string
        description:
          type: string
        active:
          type: boolean
...